# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Eko CP",
        "description": "Dokumentasi API untuk data Fase, Mapel, dan CP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/media/jenis-tema/{jenis_tema_id}/tema": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Ambil semua Tema berdasarkan Jenis Tema ID",
                "operationId": "b06f89b3f578357faecd8f434cd921c3",
                "parameters": [
                    {
                        "name": "jenis_tema_id",
                        "in": "path",
                        "description": "ID Jenis Tema",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar Tema berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "nama_tema": {
                                                        "type": "string",
                                                        "example": "Kebersihan Sebagian dari Iman"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of Tema retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/media/tema/{tema_id}/video": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Ambil semua media (video) berdasarkan Tema ID",
                "operationId": "e379371eb6efcd826abed836a79ba96a",
                "parameters": [
                    {
                        "name": "tema_id",
                        "in": "path",
                        "description": "ID Tema",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar Media berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "judul": {
                                                        "type": "string",
                                                        "example": "Menjaga Kebersihan"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://youtube.com/embed/xxxx"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "List of Media retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/media/{id}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Ambil detail media berdasarkan ID",
                "operationId": "e284861ed777bd38294a5430bca1b904",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Media (Video)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail Media berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "judul": {
                                                    "type": "string",
                                                    "example": "Video Pembelajaran Akhlak"
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "https://youtu.be/abcd1234"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Media detail retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Media tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}